rename files to pep0008 standard
