+++ scripts-python/bless_test_results
@@ -13,9 +13,9 @@ with versions of the files that you should not bless.
 from standard_script_setup import *
 
 import wait_for_tests, compare_namelists, simple_compare
-from CIME.CreateTest import NAMELIST_PHASE
+from CIME.create_test import NAMELIST_PHASE
 from CIME.utils import run_cmd, expect
-from CIME.XML.Machines import Machines
+from CIME.XML.machines import Machines
 
 import argparse, sys, os, glob, doctest, time
